Added functionality to pass values to Simulation object from command line and changed how name and mode are accessed from command line
