+++ Model/backend.py
@@ -2,6 +2,8 @@ import numpy as np
 import random as r
 import math
 import time
+import sys
+
 from numba import jit, cuda, prange
 from functools import wraps
 
@@ -878,3 +880,70 @@ def record_time(function):
         simulation.method_times[function.__name__] = end - start
 
     return wrap
+
+
+def commandline_param(flag, dtype):
+    """ Returns the value for option passed at the
+        command line.
+    """
+    # get list of command line arguments
+    args = sys.argv
+
+    # go through the arguments
+    for i in range(len(args)):
+        # if argument matches flag
+        if args[i] == flag:
+            # try to return value of
+            try:
+                return dtype(args[i + 1])
+            # otherwise raise error
+            except IndexError:
+                raise Exception(f"No value for option: {args[i]}")
+
+    # raise error if option not found
+    raise Exception(f"Option: {flag} not found")
+
+
+def template_param(path, line_number, dtype):
+    """ Gets the parameter as a string from the lines of the
+        template file. Used for Simulation instance variables.
+    """
+    # make an attribute with name as template file path and value as a list of the file lines (reduces file opening)
+    if not hasattr(template_param, path):
+        with open(path) as file:
+            template_param.path = file.readlines()
+
+    # get the right line based on the line numbers not Python indexing
+    line = template_param.path[line_number - 1]
+
+    # find the indices of the pipe characters
+    begin = line.find("|")
+    end = line.find("|", begin + 1)
+
+    # raise error if not a pair of pipe characters
+    if begin == -1 or end == -1:
+        raise Exception("Please use pipe characters to specify template file parameters. Example: | (value) |")
+
+    # return a slice of the line that is the string representation of the parameter and remove any whitespace
+    parameter = line[(begin + 1):end].strip()
+
+    # convert the parameter from string to desired data type
+    if dtype == str:
+        pass
+    elif dtype == tuple or dtype == list or dtype == dict:
+        # tuple() list() dict() will not produce desired result, use eval() instead
+        parameter = eval(parameter)
+    elif dtype == bool:
+        # handle potential inputs for booleans
+        if parameter in ["True", "true", "T", "t", "1"]:
+            parameter = True
+        elif parameter in ["False", "false", "F", "f", "0"]:
+            parameter = False
+        else:
+            raise Exception("Invalid value for bool type")
+    else:
+        # for float and int type
+        parameter = dtype(parameter)
+
+    # get the parameter by removing the pipe characters and any whitespace
+    return parameter
